Back up if wheels are stuck v3

+++ libraries/LixRobot/distance_sensor_HCSR04.h
@@ -17,12 +17,12 @@ namespace LixRobot
         {
         }
         
-        virtual unsigned int getDistance()
+        virtual long getDistance()
         {
             long distance = sensor.Ranging(CM);
             if (distance <= 0)
                 return maxDistance;
-            return (int) distance;
+            return distance;
         }
     private:
         Ultrasonic sensor;
